Refactoring code with async functions

+++ server/server.js
@@ -4,81 +4,75 @@ var _ = require('lodash');
 var express = require('express');
 var bodyParser = require('body-parser');
 
-var {mongoose} = require('./db/mongoose');
-var {Todo} = require('./models/todo');
-var {User} = require('./models/user');
-const {ObjectID} = require('mongodb');
-var {authenticate} = require('./middleware/authenticate');
+var { mongoose } = require('./db/mongoose');
+var { Todo } = require('./models/todo');
+var { User } = require('./models/user');
+const { ObjectID } = require('mongodb');
+var { authenticate } = require('./middleware/authenticate');
 
 var app = express();
 const port = process.env.PORT || 3000;
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req, res) => {
-    var todo = new Todo({
+app.post('/todos', authenticate, async (req, res) => {
+    const todo = new Todo({
         text: req.body.text,
         completed: req.body.completed,
         completedAt: req.body.completedAt,
         _creator: req.user._id
     });
-    todo.save().then((doc) => {
+    try {
+        const doc = await todo.save();
         res.send(doc);
-    }, (err) => {
+    } catch (err) {
         res.status(400).send(err);
-    });
+    }
 });
 
-app.get('/todos', authenticate, (req, res) => {
-    Todo.find({
-        _creator: req.user._id
-    }).then((todos) => {
-        res.send({todos});
-    }, (err) => {
+app.get('/todos', authenticate, async (req, res) => {
+
+    try {
+        const todos = await Todo.find({ _creator: req.user._id });
+        res.send({ todos });
+    } catch (err) {
         res.status(400).send(err);
-    });
+    }
 });
 
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos/:id', authenticate, async (req, res) => {
     if (!ObjectID.isValid(req.params.id)) {
         return res.status(404).send();
     }
-
-    Todo.findOne({
-        _id: req.params.id,
-        _creator: req.user._id
-    }).then((todo) => {
+    try {
+        const todo = await Todo.findOne({ _id: req.params.id, _creator: req.user._id });
         if (!todo) return res.status(404).send();
-        res.send({todo});
-    }).catch((err) => {
-        res.status(400).send();
-    });
-
+        res.send({ todo });
+    } catch (err) {
+        res.status(400).send(err);
+    }
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
     if (!ObjectID.isValid(req.params.id)) {
         return res.status(404).send();
     }
-    Todo.findOneAndRemove({
-        _id: req.params.id,
-        _creator: req.user._id
-    }).then((todo) => {
+    try {
+        const todo = await Todo.findOneAndRemove({ _id: req.params.id, _creator: req.user._id });
         if (!todo) return res.status(404).send();
-        res.send({todo});
-    }).catch((err) => {
-        res.status(400).send();
-    });
+        res.send({ todo });
+    } catch (err) {
+        res.status(400).send(err);
+    }
 });
 
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
+
     var id = req.params.id;
     var body = _.pick(req.body, ['text', 'completed']);
-
     if (!ObjectID.isValid(req.params.id)) {
         return res.status(404).send();
     }
-
     if (_.isBoolean(body.completed) && body.completed) {
         body.completedAt = new Date().getTime();
     } else {
@@ -86,74 +80,66 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         body.completedAt = null;
     }
 
-    Todo.findOneAndUpdate({
-        _id: id,
-        _creator: req.user._id
-    }, {$set: body}, {new: true}).then((todo) => {
+    try {
+        const todo = await Todo.findOneAndUpdate({
+            _id: id,
+            _creator: req.user._id
+        }, { $set: body }, { new: true });
         if (!todo) return res.status(404).send();
-        res.send({todo});
-    }).catch((err) => {
-        res.status(400).send();
-    });
-
-});
-
-
-app.get('/users', (req, res) => {
-    User.find({}).then((users) => {
-        res.send(
-            {
-                'count': users.length,
-                users
-            });
-    }, (err) => {
-        console.log(err);
-    });
-});
-
-app.post('/users', (req, res) => {
-    var body = _.pick(req.body, ['email', 'password']);
-
-    var user = new User(body);
-    user.save().then((user) => {
-
-        return user.generateAuthToken();
-
-    }).then((token) => {
-        res.header('x-auth', token).send(user);
-    }).catch((err) => {
+        res.send({ todo });
+    } catch (err) {
         res.status(400).send(err);
-    });
+    }
+});
 
+
+app.get('/users', async (req, res) => {
+    try {
+        const users = await User.find({});
+        res.send({ 'count': users.length, users });
+    } catch (err) {
+        res.status(400).send(err);
+    }
+});
+
+app.post('/users', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = new User(body);
+        const resultUser = await user.save();
+        const token = await resultUser.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (err) {
+        res.status(400).send(err);
+    }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-    var body = _.pick(req.body, ['email', 'password']);
-
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });
-    }).catch((err) => {
-        res.status(400).send();
-    });
-
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (err) {
+        res.status(400).send(err);
+    }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token);
         res.status(200).send();
-    }, () => {
-        res.status(400).send();
-    });
+    } catch (err) {
+        res.status(400).send(err);
+    }
 });
 
 app.listen(3000, () => {
     console.log(`Started on port ${port}`);
 });
 
-module.exports = {app};
+module.exports = { app };
